Add constants to allow 22MB and 33MB data rate
The update of the driver to the version 'Realtek v5.5.2' introduce
a check of the DR broadcasted in the WLAN_EID_SUPP_RATES IE.
As a conseconsed AP with 22MB or 33MB Data Rate are filtered out.
The check has been added in to function collect_bss_info of rtw_mlne_ext.c
and performed in the function rtw_validate_value of rtw_wlan_util.c

Adding these constants will allow to make the AP available again.
Even if I did not observe any issue, I don't known if the driver is
able to support such data rate.

Anyway this patch permits to connect again on "Freebox" AP, the modem
provided by French ISP "Free". Issue #64